refactor: . Move showDismissibleNotice() to DismissibleNotices.ts
